+++ WITH_Server/WITH_Server/Stats.h
@@ -14,7 +14,7 @@ struct BaseAttribute : Component
 	int power{ 10 };
 	double attackSpeed{ 1.0 };
 	int defense{ 10 };
-	int moveSpeed{ 2 };
+	double moveSpeed{ 2.0 };
 };
 
 struct FinalVital : Component
@@ -30,7 +30,7 @@ struct FinalAttribute : Component
 	int power{ 10 };
 	double attackSpeed{ 1.0 };
 	int defense{ 10 };
-	int moveSpeed{ 2 };
+	double moveSpeed{ 2.0 };
 	bool dirty{ false };
 };
